Struct hexeng::Frame

pub struct Frame<V: Scalar, T>(_, _);

An object representing a point within a specified frame.

The frame is included in the type by the parameter T, which should be one of TAxial, TWorld or TView.

Methods

impl<V: Scalar, T> Frame<V, T>

pub fn as_slice(&self) -> &[V]

pub fn as_array(self) -> [V; 2]